requirements.txt
pyyaml==5.3.1
psycopg2==2.8.5
pyowm==2.10.0
freezegun==1.0.0
RPi.GPIO==0.7.0